+++ src/sc_client/session.py
@@ -17,11 +17,11 @@ import websocket
 from sc_client.client._executor import Executor
 from sc_client.constants import common
 from sc_client.constants.common import ClientCommand
-from sc_client.constants.numeric import LOGGING_MAX_SIZE, SERVER_ANSWER_CHECK_TIME, SERVER_ESTABLISH_CONNECTION_TIME
+from sc_client.constants.numeric import LOGGING_MAX_SIZE, SERVER_ANSWER_CHECK_TIME, SERVER_ESTABLISH_CONNECTION_TIME, SERVER_RECONNECTION_TIME
 from sc_client.models import Response, ScAddr, ScEvent
 
 logger = logging.getLogger(__name__)
-
+reconnection_enable = False
 
 class _ScClientSession:
     lock_instance = threading.Lock()
@@ -71,11 +71,21 @@ def is_connection_established() -> bool:
     return bool(_ScClientSession.ws_app)
 
 
-def set_connection(url: str) -> None:
+def set_connection(url: str, autoreconnection = False, reconnection_time = SERVER_RECONNECTION_TIME) -> None:
+    reconnection_enable = autoreconnection
+    if (reconnection_time > 1800): #30 min
+        reconnection_time = SERVER_RECONNECTION_TIME
     if not is_connection_established():
         establish_connection(url)
-    if not is_connection_established():
-        raise ConnectionRefusedError
+
+    def run_in_thread(url: str):
+        while (reconnection_enable):
+            if not is_connection_established():
+                time.sleep(reconnection_time)
+                establish_connection(url)
+
+    reconnection_thread = threading.Thread(target=run_in_thread, args=(url,), name="reconnection-thread")
+    reconnection_thread.start()
 
 
 def establish_connection(url) -> None:
@@ -96,6 +106,7 @@ def establish_connection(url) -> None:
 
 def close_connection() -> None:
     try:
+        reconnection_enable = False
         _ScClientSession.ws_app.close()
     except AttributeError:
         pass